`Directory` should not have `init_dir()` #206
