+++ src/storage/directory.rs
@@ -1,4 +1,3 @@
-use std::fs::create_dir_all;
 use std::path::{Path, PathBuf};
 use crate::name::Name;
 
@@ -65,48 +64,5 @@ pub trait Directory: Name {
         self.parent_dir()
             .expect("No parent directory has been set")
             .join(self.dir_name())
-    }
-
-    /// Builder method that creates dedicated directory
-    ///
-    /// If directory already exists, then this method silently fails.
-    ///
-    /// # Panics
-    ///
-    /// This method panics if an error occurs when creating directory (other than directory
-    /// already existing). This could happen if write permissions are misconfigured.
-    ///
-    /// # Returns
-    ///
-    /// Ownership of `Self`, allowing method chaining.
-    fn init_dir(self) -> Self
-        where
-            Self: Sized
-    {
-        self.init_dir_ref();
-        self
-    }
-
-    /// Method that creates dedicated directory without taking ownership of `self`
-    ///
-    /// If directory already exists, then this method silently fails.
-    ///
-    /// # Panics
-    ///
-    /// This method panics if an error occurs when creating directory (other than directory
-    /// already existing). This could happen if write permissions are misconfigured.
-    ///
-    /// # Returns
-    ///
-    /// Immutable reference of `Self`, allowing method chaining.
-    fn init_dir_ref(&self) -> &Self {
-        let path = self.full_path();
-        match path.exists() {
-            true => (),
-            false => {
-                create_dir_all(path).expect("Could not create dedicated directory");
-            }
-        };
-        self
     }
 }
